Bulk upload and create new user Fix (#35)
* Email config change

* Removed location cache

* Heroku Test Env: redis remove update

* Heroku Test Env: redis remove update in properties and config

* Heroku Test Env: redis remove update in properties and config update

* Forget password fix

* Bulk upload and create new user Fix

* Production Fixes

1. Multi Location
2. Multi Group in create user
3. Redis cache based login privilege check
4. Email content and validations
5. Attendance supervisor list for mobile view
6. Profile endpoint fix
7. Report group response fix

* Test Environment Fixes

1. In create User method included inactive user handing flow and alert
2. Allow one time edit - paid leave based scenario
3. Fixed group response to Array List in getAllUser method
4. Add Location duplicate fix

* Feature test environment (#36)

* Heroku Prod- Forget password mail fix

* Heroku Prod-create user log check

* Bulk upload and create new user Fix

* Production Fixes

1. Multi Location
2. Multi Group in create user
3. Redis cache based login privilege check
4. Email content and validations
5. Attendance supervisor list for mobile view
6. Profile endpoint fix
7. Report group response fix

* Test Environment Fixes

1. In create User method included inactive user handing flow and alert
2. Allow one time edit - paid leave based scenario
3. Fixed group response to Array List in getAllUser method
4. Add Location duplicate fix

---------


* Bug Fix

Redis

* Privilege Cache

Organization based endpoint fix
privilege based cache fix

* Update application property

* Update redisScheduler

* Test bug fix for production

1. Fixed bulk upload
2. Created new email template
3. Fixed secondary details updation method
4. Fixed secondary mobile number constrain

---------

+++ src/main/java/com/uniq/tms/tms_microservice/repository/TeamRepository.java
@@ -47,7 +47,6 @@ public interface TeamRepository extends JpaRepository<GroupEntity, Long> {
             u.email,
             u.date_of_joining,
             r.name AS role_name,
-            u.location_id,
             u.active
         FROM users u
         LEFT JOIN role r ON u.role_id = r.role_id
@@ -111,14 +110,14 @@ public interface TeamRepository extends JpaRepository<GroupEntity, Long> {
     );
 
     @Query(value = """
-        SELECT 
+        SELECT
             g.group_id AS groupId,
             g.group_name AS groupName
-        FROM 
+        FROM
             group_table g
-        JOIN 
+        JOIN
             user_group ug ON g.group_id = ug.group_id
-        WHERE 
+        WHERE
             ug.user_id = :userId
             AND g.organization_id = :orgId
             AND ug.type = 'Supervisor'
@@ -133,8 +132,9 @@ public interface TeamRepository extends JpaRepository<GroupEntity, Long> {
     @Query(value = """
         DELETE FROM group_table
         WHERE group_id = :groupId
+        AND organization_id = :orgId
         """, nativeQuery = true)
-    void deleteGroupById(@Param("groupId") Long groupId);
+    void deleteGroupById(@Param("groupId") Long groupId, @Param("orgId") Long orgId);
 
     @Query(value = """
         SELECT 
@@ -152,4 +152,7 @@ public interface TeamRepository extends JpaRepository<GroupEntity, Long> {
     @Query(value = "SELECT group_name, group_id FROM group_table", nativeQuery = true)
     List<Object[]> findGroupNameIdMappings();
 
+    List<GroupEntity> findAllByOrganizationEntity_OrganizationId(Long orgId);
+
+    boolean existsByGroupIdAndOrganizationEntity_OrganizationId(Long groupId, Long orgId);
 }
